Use size_t instead of off_t when handling memory addresses. This prevents dmidecode from crashing on some systems with more than 2 GB of memory.

+++ util.c
@@ -95,7 +95,7 @@ int checksum(const u8 *buf, size_t len)
  * Copy a physical memory chunk into a memory buffer.
  * This function allocates memory.
  */
-void *mem_chunk(off_t base, off_t len, const char *devmem)
+void *mem_chunk(size_t base, size_t len, const char *devmem)
 {
 	void *p;
 	int fd;
