+++ util.h
@@ -3,4 +3,4 @@
 #include "types.h"
 
 int checksum(const u8 *buf, size_t len);
-void *mem_chunk(off_t base, off_t len, const char *devmem);
+void *mem_chunk(size_t base, size_t len, const char *devmem);
